requirements.txt
assemblyai
anthropic
openai
yt-dlp
python-dotenv
